standardized user home page event table
