+++ src/screens/Events/index.jsx
@@ -6,8 +6,7 @@ const EventManagerSelector = () => {
   const {
     data: { user },
   } = useSession();
-
-  return <EventManager user={user} role={user.role} />
+  return <EventManager user={user} role={user.role} isHomePage={false} />
 };
 
 export default EventManagerSelector;
