+++ src/screens/Homepage/index.jsx
@@ -1,7 +1,7 @@
 import { useSession } from "next-auth/react";
 import React from "react";
-import UserEventManager from "./User";
-import EventManager from "../Events";
+// import UserEventManager from "./User";
+import EventManager from "../Events/EventManager";
 
 const EventManagerSelector = () => {
   const {
@@ -9,9 +9,9 @@ const EventManagerSelector = () => {
   } = useSession();
 
   if (user.role !== "volunteer") {
-    return <EventManager user={user} role={user.role} />;
+    return <EventManager user={user} role={user.role} isHomePage={false} />;
   } else {
-    return <UserEventManager user={user} />;
+    return <EventManager user={user} role={user.role} isHomePage={true} />;
   }
 };
